% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_data.R
\name{Generate_data}
\alias{Generate_data}
\title{Generate_data}
\usage{
Generate_data(truth.beta, p, n, truth.sigma, rho, correlation = c("block", "all"),
NumOfBlock)
}
\arguments{
\item{truth.beta}{A vector of active beta's (s x 1, with s being the number of active coordinates).}

\item{p}{The number of covariates.}

\item{n}{The number of observations.}

\item{truth.sigma}{Noise standard deviation.}

\item{rho}{Correlation Coefficient.}

\item{correlation}{Correlation structure. Correlation = "block" means predictors are grouped into equi-size blocks where each block contains one active predictor, and the within-block correlation coefficient is rho; predictors in different blocks are mutually independent. Correlation = "all" means all predictors are equi-correlated with coefficient rho.}

\item{NumOfBlock}{Number of blocks, used only when correlation = 'block'.}
}
\value{
A list, including vector 'y' (n x 1), matrix 'X' (n x p), vector 'beta' (p x 1).
}
\description{
This function generates data (X,y) with specified correlation and noise standard deviation.
}
